The first working option

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -12,46 +12,31 @@ import java.util.List;
 @Component
 public class UserDaoImpl implements UserDao {
     @PersistenceContext
-    private final EntityManager entityManager;
+    private EntityManager entityManager;
 
-    @Autowired
-    public UserDaoImpl(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
     @Override
     public List<User> getAllUsers() {
-        return entityManager.createQuery("select u from User u", User.class).getResultList();
+        return entityManager.createQuery("FROM User", User.class).getResultList();
     }
+
     @Override
-    public void save(User user) {
+    public void addUser(User user) {
         entityManager.persist(user);
     }
+
     @Override
-    public User show(int id) {
-        TypedQuery<User> query = entityManager.createQuery("select u from User u where u.id = :id", User.class);
-        query.setParameter("id", id);
-        return query.getSingleResult();
-    }
-    @Override
-    public void update(int id, User updateUser) {
-        User user = show(id);
-        user.setName(updateUser.getName());
-        user.setSurname(updateUser.getSurname());
-        user.setAge(updateUser.getAge());
+    public void updateUser(User user) {
         entityManager.merge(user);
     }
+
     @Override
-    public void delete(int id) {
-        User user = show(id);
-        entityManager.remove(user);
+    public void deleteUser(int id) {
+        User user = getUserById(id);
+        if (user != null) entityManager.remove(user);
     }
+
     @Override
-    public User isExistById(User user) {
-        if(entityManager.contains(user)) {
-            entityManager.remove(user);
-        } else {
-            entityManager.remove(entityManager.merge(user));
-        }
-        return user;
+    public User getUserById(int id) {
+        return entityManager.find(User.class, id);
     }
 }
